+++ collective/googlenews/config.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 PROJECTNAME = 'collective.googlenews'
+GUARD_EXPRESSION = 'here/@@googlenews-helper/publish_transition_allowed'