+++ collective/googlenews/testing.py
@@ -33,6 +33,26 @@ IMAGES = [
 ]
 
 
+def _setup_content(n, in_the_past=False):
+    """Create and publish dummy content with random ids."""
+    from DateTime import DateTime
+    from uuid import uuid1
+    context = api.portal.get()
+    for i in range(0, n):
+        with api.env.adopt_roles(['Manager']):
+            id_ = uuid1().hex[:8]  # generate random id
+            obj = api.content.create(
+                context, 'Dexterity Item', id_, standout_journalism=True)
+            api.content.transition(obj=obj, transition='publish')
+            if in_the_past:
+                obj.effective_date = DateTime('01/01/2000')
+            else:
+                # XXX: api.content.transition doesn't set effective_date
+                #      https://github.com/plone/plone.api/issues/343
+                obj.effective_date = DateTime()
+            obj.reindexObject()
+
+
 class Fixture(PloneSandboxLayer):
 
     defaultBases = (PLONE_FIXTURE,)
@@ -59,44 +79,27 @@ class Fixture(PloneSandboxLayer):
             shutil.copy2(img_path, '/tmp')
 
 
+FIXTURE = Fixture()
+
+INTEGRATION_TESTING = IntegrationTesting(
+    bases=(FIXTURE,), name='collective.googlenews:Integration')
+
+FUNCTIONAL_TESTING = FunctionalTesting(
+    bases=(FIXTURE,), name='collective.googlenews:Functional')
+
+
 class RobotFixture(Fixture):
 
     """This fixture adds content to test the standout journalism validator."""
 
     def setUpPloneSite(self, portal):
         super(RobotFixture, self).setUpPloneSite(portal)
-        from DateTime import DateTime
-        from plone import api
-
-        with api.env.adopt_roles(['Manager']):
-            folder = api.content.create(portal, 'Folder', 'folder')
-
-            # we will create and publish 7 news articles
-            for i in range(0, 7):
-                obj = api.content.create(
-                    folder, 'Dexterity Item', str(i), standout_journalism=True)
-                api.content.transition(obj, transition='publish')
-                obj.setEffectiveDate(DateTime())
-                obj.reindexObject()
-
-        # one of them will be set as published one week in the past
-        folder['0'].setEffectiveDate(DateTime() - 8)
-        folder['0'].reindexObject()
-
-FIXTURE = Fixture()
-
-INTEGRATION_TESTING = IntegrationTesting(
-    bases=(FIXTURE,),
-    name='collective.googlenews:Integration',
-)
-
-FUNCTIONAL_TESTING = FunctionalTesting(
-    bases=(FIXTURE,),
-    name='collective.googlenews:Functional',
-)
+        # create and publish 6 items marked as standout journalism
+        _setup_content(6)
 
 ROBOT_FIXTURE = RobotFixture()
 
 ROBOT_TESTING = FunctionalTesting(
     bases=(ROBOT_FIXTURE, AUTOLOGIN_LIBRARY_FIXTURE, z2.ZSERVER_FIXTURE),
-    name='collective.googlenews:Robot')
+    name='collective.googlenews:Robot',
+)