+++ collective/googlenews/tests/test_behaviors.py
@@ -1,6 +1,11 @@
 # -*- coding: utf-8 -*-
+from collective.googlenews.testing import _setup_content
 from collective.googlenews.testing import INTEGRATION_TESTING
+from collective.googlenews.utils import get_current_standout_journalism
 from plone import api
+from plone.api.exc import InvalidParameterError
+from plone.app.testing import setRoles
+from plone.app.testing import TEST_USER_ID
 
 import unittest
 
@@ -11,11 +16,9 @@ class BehaviorsTestCase(unittest.TestCase):
 
     def setUp(self):
         self.portal = self.layer['portal']
-
-        with api.env.adopt_roles(['Manager']):
-            self.folder = api.content.create(self.portal, 'Folder', 'folder')
-
-        self.dummy1 = api.content.create(self.folder, 'Dexterity Item', 'd1')
+        self.request = self.layer['request']
+        setRoles(self.portal, TEST_USER_ID, ['Manager'])
+        self.dummy1 = api.content.create(self.portal, 'Dexterity Item', 'd1')
 
     def test_googlenews_behavior(self):
         from collective.googlenews.behaviors.interfaces import IGoogleNews
@@ -31,3 +34,23 @@ class BehaviorsTestCase(unittest.TestCase):
 
         self.assertTrue(self.dummy1.standout_journalism)
         self.assertEqual(len(self.dummy1.news_keywords), 3)
+
+    def test_standout_journalism_validation(self):
+        # create and publish 7 items in the past
+        _setup_content(7, in_the_past=True)
+        results = get_current_standout_journalism()
+        self.assertEqual(len(results), 0)
+
+        # create and publish 7 items
+        _setup_content(7)
+        results = get_current_standout_journalism()
+        self.assertEqual(len(results), 7)
+
+        # we can publish a new item as long as it's not marked as standout
+        obj = api.content.create(self.portal, 'Dexterity Item', 'foo')
+        api.content.transition(obj=obj, transition='publish')
+
+        obj = api.content.create(
+            self.portal, 'Dexterity Item', 'bar', standout_journalism=True)
+        with self.assertRaises(InvalidParameterError):
+            api.content.transition(obj=obj, transition='publish')