+++ collective/googlenews/tests/test_setup.py
@@ -2,6 +2,7 @@
 from collective.googlenews.config import PROJECTNAME
 from collective.googlenews.interfaces import IGoogleNewsLayer
 from collective.googlenews.testing import INTEGRATION_TESTING
+from collective.googlenews.utils import get_workflows_with_publish_transition
 from plone import api
 from plone.browserlayer.utils import registered_layers
 
@@ -22,6 +23,12 @@ class InstallTestCase(unittest.TestCase):
     def test_browserlayer_installed(self):
         self.assertIn(IGoogleNewsLayer, registered_layers())
 
+    def test_workflow_guards_installed(self):
+        from collective.googlenews.config import GUARD_EXPRESSION
+        for wf in get_workflows_with_publish_transition():
+            guard = wf.transitions['publish'].guard
+            self.assertEqual(guard.getExprText(), GUARD_EXPRESSION)
+
 
 class UninstallTestCase(unittest.TestCase):
 
@@ -38,3 +45,8 @@ class UninstallTestCase(unittest.TestCase):
 
     def test_browserlayer_removed(self):
         self.assertNotIn(IGoogleNewsLayer, registered_layers())
+
+    def test_workflow_guards_removed(self):
+        for wf in get_workflows_with_publish_transition():
+            guard = wf.transitions['publish'].guard
+            self.assertEqual(guard.getExprText(), '')